a lot of new stuff.. layouts, presentation layer, some refactoring
this is a bad commit :( sorry

+++ app/src/main/java/com/vp/weatherapp/data/remote/ForecastResponse.kt
@@ -34,8 +34,8 @@ data class City(val id: Long, val name: String, val coord: Coordinates,
 data class Coordinates(val lon: Float, val lat: Float)
 
 
-data class Temperature(val day: Float, val min: Float, val max: Float,
-                       val night: Float, val eve: Float, val morn: Float)
+data class Temperature(val day: Double, val min: Double, val max: Double,
+                       val night: Double, val eve: Double, val morn: Double)
 
 data class Weather(val id: Long, val main: String, val description: String,
                    val icon: String)
